Decided on an approach to handling the 2 menus (this would not work if there were more than 2 menus but could in theory be expanded.
Currently the LibraryHandler will have 2 nextAction() commands with each being the relevant menu for whether it's a record or a system. Ideally this would work for if we had more menus but in reality we'd have to abstract out the Menu bit into an interface that the LibraryHandler would implement.

This works for now...